+++ api/api-default-impl/src/main/java/net/suqatri/redicloud/api/impl/configuration/impl/PlayerConfiguration.java
@@ -1,4 +1,4 @@
-package net.suqatri.redicloud.api.impl.player;
+package net.suqatri.redicloud.api.impl.configuration.impl;
 
 import lombok.Data;
 import net.suqatri.redicloud.api.impl.configuration.Configuration;
